% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIncidence.R
\name{constraint.function}
\alias{constraint.function}
\title{Constraint function}
\usage{
constraint.function(x, lower, upper, ...)
}
\value{
Constraint function for upper / lower bounds
}
\description{
Computes the constraint function for use in constrOptim.nl (hin parameter)
}
\keyword{internal}
