% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIncidence.R
\name{constraint.jacobian.create}
\alias{constraint.jacobian.create}
\title{Creates the Jacobian of the constraint function}
\usage{
constraint.jacobian.create(x, variables = length(x))
}
\value{
Jacobian of constraint function
}
\description{
Creates Jacobian of the constraint function for use in constrOptim.nl
   (hin.jac parameter)
}
\keyword{internal}
