% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIncidence.R
\name{constraint.jacobian.function}
\alias{constraint.jacobian.function}
\title{Jacobian of the constraint function}
\usage{
constraint.jacobian.function(constraint.jacobian, ...)
}
\value{
Jacobian of constraint function
}
\description{
Returns the previously-computed Jacobian of the constraint function for
   use in constrOptim.nl (hin.jac parameter)
}
\keyword{internal}
