% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flumodelsutil_data}
\alias{flumodelsutil_data}
\title{flumodelsutil_data}
\format{
An object of class \code{list} of length 7.
}
\usage{
data(flumodelsutil_data)
}
\description{
flumodelsutil_data
}
\keyword{datasets}
