% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeContactMatrix.R
\name{makeContactMatrix}
\alias{makeContactMatrix}
\title{Make contact matrix}
\usage{
makeContactMatrix(ages, originalContactMatrix = fitflumodels::POLYMOD.matrix,
  originalContactMatrixAges = fitflumodels::POLYMOD.age.ranges,
  originalPopulationFractions = fitflumodels::population.UK)
}
\arguments{
\item{ages}{Vector of ages. Each element represents the upper range of an age range. The lowest bound is presumed to be zero.
If the oldest age does not reach the end of the population range, an additional element is added to span the full range.
The final age bracket cannot start after the final bracket of originalContactMatrixAges.}

\item{originalContactMatrix}{Contact matrix to serve as the basis for extrapolation. Defaults to UK POLYMOD.
This must be a square matrix.}

\item{originalContactMatrixAges}{Age ranges associated with originalContactMatrix. Defaults to UK population in 5-year bins, with
the final bracket starting at age 70. The format is a data frame with columns representing the age brackets and having two rows:
AgeStart and AgeEnd. The first column has ageStart of 0. The last column has ageEnd of NA.}

\item{originalPopulationFractions}{Vector of age fractions associated with originalContactMatrixAges.
Length must equal the dimension of originalContactMatrix. The vector will be normalized, so it can represent population fractions
or population in each age bin.}
}
\value{
A contact matrix.
}
\description{
Rebins an arbitrary contact matrix to a new group of population ranges. The default behavior is to use the UK POLYMOD data.
}
\author{
Jason Asher <jason.m.asher@gmail.com>
}
