% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fitIncidence}
\alias{plot.fitIncidence}
\title{Plot fitIncidence}
\usage{
\method{plot}{fitIncidence}(x, ...)
}
\arguments{
\item{x}{fitIncidence object created by fitIncidence function}
}
\description{
Creates a plot comparing the fit and original incidence.
}
\keyword{internal}
