% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYMOD.age.ranges.R
\docType{data}
\name{POLYMOD.age.ranges}
\alias{POLYMOD.age.ranges}
\title{Age ranges associated with POLYMOD contact matrix for Great Britain.}
\format{a \code{data.frame}, where each column corresponds to an age range. The
column names denote the age ranges. The two rows are AgeStart and AgeEnd.
The given age range corresponds to individuals in the range [AgeStart, AgeEnd].
A value of NA for AgeEnd corresponds to no upper bound, so [70,NA] corresponds
to all individuals aged 70 or more.}
\source{
Mossong J, Hens N, Jit M, Beutels P, Auranen K, et al. (2008)
Social Contacts and Mixing Patterns Relevant to the Spread of Infectious Diseases.
PLOS Medicine 5(3): e74. https://doi.org/10.1371/journal.pmed.0050074
}
\usage{
POLYMOD.age.ranges
}
\description{
Data frame that provides information on the age ranges associated with
the POLYMOD data.
}
\keyword{datasets}
