% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYMOD.matrix.R
\docType{data}
\name{POLYMOD.matrix}
\alias{POLYMOD.matrix}
\title{POLYMOD contact matrix for Great Britain.}
\format{a \code{double} matrix, matrix where k_ij is proportional to the observed number of contacts
(both physical and nonphysical) that a respondent in age band j makes with other
individuals in age band i. The names of both columns and rows denote the
age ranges.}
\source{
Mossong J, Hens N, Jit M, Beutels P, Auranen K, et al. (2008)
Social Contacts and Mixing Patterns Relevant to the Spread of Infectious Diseases.
PLOS Medicine 5(3): e74. https://doi.org/10.1371/journal.pmed.0050074
}
\usage{
POLYMOD.matrix
}
\description{
Scaled contact matrix suitable for use in disease transmission modeling.
}
\keyword{datasets}
