% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.UK.R
\docType{data}
\name{population.UK}
\alias{population.UK}
\title{Population of the UK for use with POLYMOD data.}
\format{a \code{integer} array, one entry per age range.
The \code{names} denote the age ranges.}
\source{
???
}
\usage{
population.UK
}
\description{
This data set is the population of the UK by age range.
It is designed to aid in rebinning the POLYMOD contact matrix.
}
\keyword{datasets}
