% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.US.R
\docType{data}
\name{population.US}
\alias{population.US}
\title{Population data for US from 2015.}
\format{a \code{data.frame} with many columns. Those of interest for this package are:
\itemize{
\item YEAR: Which year the data corresponds to. Currently only 2015 data is provided.
\item AGE: Age of the population of interest. Age of X corresponds to [X,X+1),
so 0 means infants under one year of age. Age of 999 corresponds to the entire population.
\item TOT_POP: Total population in the given age bin (or entire US population if AGE == 999).
}}
\source{
US Census
}
\usage{
population.US
}
\description{
Data from US Census 2015 estimate on US population by single age bins.
}
\keyword{datasets}
