% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleIncidence.R
\name{rescaleIncidence}
\alias{rescaleIncidence}
\title{Rescale incidence}
\usage{
rescaleIncidence(incidence, totalInfections, weeksOfBackground = 3)
}
\arguments{
\item{incidence}{Weekly incidence}

\item{totalInfections}{Total number of infections. It can represent the total number of infections or 
the fraction of the population that became infected (totalInfections < 1).}

\item{weeksOfBackground}{Weeks of data at start and end of the data to presume reflect "background" incidence and to reduce the
incidence curve by. Default to 3.}
}
\value{
A vector of weekly incidence. If totalInfections represents a fraction of the population (< 1), then 
   this represents a rate of infection.
}
\description{
Rescales incidence to match a given total number of infections. Can adjust to reduce incidence by a presumed background
   rate.
}
\author{
Matt Clay <clay.matt@gmail.com>
}
\keyword{internal}
