% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIncidence.R
\name{SEIR.penalty.function}
\alias{SEIR.penalty.function}
\title{Penalty function for SEIR-type models}
\usage{
SEIR.penalty.function(model, incidence, attackRatesToMatch, populationFractions,
  population = 1, fractionSymptomatic = 0.5, seedWeeksBeforeIncidence = 0,
  weightIncidenceTimeseries = 1, weightAttackRates = 1)
}
\arguments{
\item{model}{SEIR-type model}

\item{incidence}{Array of weekly incidence}

\item{attackRatesToMatch}{Attack rates for population groups}

\item{populationFractions}{Population fractions}

\item{population}{Total population, Default: 1}

\item{fractionSymptomatic}{The fraction of cases that are symptomatic. Defaults to 0.5}

\item{seedWeeksBeforeIncidence}{Days before the start of the supplied incidence curve to seed initial infections.}

\item{weightIncidenceTimeseries}{What relative weight to the penalty function should the incidence timeseries receive? Defaults to 1.}

\item{weightAttackRates}{What relative weight to the penalty function should the age-specific attack rates receive? Defaults to 1.}
}
\value{
Total penalty
}
\description{
Computes the penalty compared to preexisting incidence.
}
\keyword{internal}
